Implement Save/Restore Cookies
The authentication cookies which are returned during the login process
should be saved and reused rather than calling the login endpoint each
time a call needs to be made. In order to do this the application
consuming the API needs a way for the cookie values to be accessed.
These functions allow for the cookie values to be returned to the
calling code to be saved and then supplied as necessary at a later
point.