Rename Main Client, User/Pass on Options, Client Interface Extracted
The main client has had an interface extracted and now that is what gets
returned for use. This interface looked strange being
"IiRacingDataClient" due to the two i's at the start. This led to a
rename of the main client class to remove the "iRacing" name.

The login method was made internal and triggered on any method call when
the "IsLoggedIn" property is false.