Use TimeSpan for SubsessionChartLap Interval
The "SubsessionChartLap" value doesn't use the normal ten-thousandth of
a second number for the interval. The interval is a combination of two
fields, one with a numerical value and the other with the units of that
value. The example available to the project currently uses "ms" as the
unit which indicates it is in "milliseconds". This code handles that and
will convert appropriately into a "TimeSpan" on a separate property.

Other units will throw an "InvalidOperationException" for now.