Provide "Race Week Number" Convenience Property
The iRacing Data API operates using zero-based race week numbers. For
most display or processing purposes people will prefer to work with a
one-based count.

This change adds a convenience method which can be used to directly
display or work with the actual one-based week number instead of the
zero-based "index".

Fixes: #52