Properly Handle Login Failures
The code was relying on a non-success status code to verify login. That
doesn't work as iRacing will return a "200" code with a JSON body that
indicates the authentication failed. These changes check both the HTTP
status and the JSON body to ensure the authentication was successful.
If not, an exception is thrown.