Rework Builds
Restructure the builds to be based around the same build script but
include it. This means we can do away with mulitple workflows executing
for a single action (e.g. one "build" workflow and one "create release"
workflow) and instead a single workflow with multiple jobs can be
executed for a single task.

This means a continuous integration build is one workflow with build,
test report, and deployment of pre-release, instead of two workflows as
it was before.