Use Single Successful Login Response
Previously the successful login response was duplicated across the
tests. This meant it was impossible to simply stack responses from two
different tests so that two endpoints could return the captured data.

Now that the login response is separate for almost all the response
packs it can be added based on a flag. So the first call to add data
responses can include the login but later calls, to enable other
methods, can omit the login.