Implement Locking for Authentication
It is possible on a particular DataClient instance to call multiple
methods and have each method trigger the authentication process. This
leads to the login calls being made multiple times, wasting API calls.

This change places a lock such that the login process will be called
once regardless of how many methods attempt to access it at the same
time.

Fixes: #185