Use Cookies Even Without "RestoreCookies"
Previously the code only checked the cookie container for relevant
cookies if the "RestoreCookies" delegate was configured. This means that
if the cookie container was populated some other way then the login
process would be called again.

Move the check for cookies which might be existing, valid authentication
outside the check for "RestoreCookies" delegate to take advantage of any
data the collection already contains.