Base Member Info Integration Test on Customer ID
iRacing have changed the result of the Member Info call to exclude the
email address for privacy reasons. So we can no longer use this to check
our result is expected. Instead, use my customer identifier.

Remove the class that imitated the previous security masking as it is no
longer necessary.